# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalalPilot Orchestration API
  version: 0.1.0
  description: |
    Kontrak antara skill OpenClaw `halalpilot` (pemanggil) dan layanan Express+SQLite (pelaksana).
    Prinsip: agent memutuskan bahasa & alur percakapan; API memutuskan hukum (rules/*.yaml), status, dan jadwal.
    Semua endpoint di 127.0.0.1:3000, header `Authorization: Bearer ${HALALPILOT_API_TOKEN}`.
    Endpoint yang mengubah state menerima `Idempotency-Key` (disarankan skill mengisi `<aksi>:<umk>:<hash payload>`).
servers:
  - url: http://127.0.0.1:3000/api/v1
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  schemas:
...
    UmkSummary:
      type: object
      properties:
        id: { type: integer }
        kode: { type: string, example: UMK-017 }
        nama_usaha: { type: string }
        status: { type: string }
        skor_kesiapan: { type: integer }
        jalur_terakhir: { type: string, nullable: true }
        dokumen_kurang: { type: array, items: { type: string } }
...
    Decision:
      type: object
      properties:
        jalur: { type: string, enum: [SELF_DECLARE_SIAP, SELF_DECLARE_KURANG_DOKUMEN, REGULER, TIDAK_LAYAK] }
        skor_kesiapan: { type: integer }
        alasan:
          type: array
          items:
            type: object
            properties:
              rule_id: { type: string }
              hasil: { type: string, enum: [lolos, gagal, butuh_dokumen] }
              pesan_umk: { type: string }
        dokumen_diminta: { type: array, items: { type: string } }
        rules_version: { type: string }
...
paths:
...
  /umk:
    # GET mengembalikan per UMK: id, kode, nama_usaha, status, jalur, skor_kesiapan, dokumen_kurang, dokumen_total, dokumen_diterima, eskalasi, produk, last_activity_at, hari_diam, consent (untuk papan dashboard)
    get:
      summary: Daftar UMK satu koperasi untuk papan dashboard/pendamping (status, skor, jalur, dokumen kurang, eskalasi). Bukan untuk aktor UMK.
      parameters: [{ name: koperasi_id, in: query, schema: { type: integer, default: 1 } }]
      responses: { '200': { description: "{koperasi_id, umk:[{id,kode,nama_usaha,status,jalur,skor_kesiapan,dokumen_kurang,eskalasi,produk,updated_at}]}" } }
    post:
      summary: Daftarkan UMK baru (intake awal) + catat consent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [telegram_id, nama_usaha, consent]
              properties:
                telegram_id: { type: string }
                nama_usaha: { type: string }
                nib: { type: string, nullable: true }
                consent: { type: boolean, description: "UMK menyetujui pemrosesan data usaha untuk penyiapan berkas halal" }
      responses:
        '201': { description: Dibuat, content: { application/json: { schema: { $ref: '#/components/schemas/UmkSummary' } } } }
  /umk/{id}:
    get:
      summary: Ringkasan UMK, produk, dokumen kurang, keputusan terakhir
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      responses: { '200': { description: OK } }
    patch:
      summary: Lengkapi profil usaha (NIB, KBLI, alamat, omzet, fasilitas, peralatan, penyelia)
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: OK } }
    delete:
      summary: Hapus seluruh data UMK (hak subjek data, UU PDP) — mengembalikan bukti penghapusan
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      responses: { '200': { description: "{deleted:true, hash_bukti}" } }
  /umk/{id}/products:
    post:
      summary: Tambah produk
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [nama]
              properties:
                nama: { type: string }
                jenis: { type: string }
                kbli: { type: string, nullable: true }
      responses: { '201': { description: Dibuat } }
...
  /umk/{id}/evaluate:
    post:
      summary: Jalankan mesin keputusan (rules/eligibility.yaml + oss-check.yaml). Membuat baris decision & document_req.
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Decision' } } } }
  /umk/{id}/documents/{kode}:
    put:
      summary: Terima dokumen (media_id) atau tandai dihasilkan; membatalkan chase_task terkait
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }
        - { name: kode, in: path, required: true, schema: { type: string } }
      requestBody: { content: { application/json: { schema: { type: object, properties: { media_id: { type: integer }, catatan: { type: string } } } } } }
      responses: { '200': { description: OK } }
  /umk/{id}/supplier-certs:
    post:
      summary: Daftarkan sertifikat pemasok; API memeriksa ke registry mock (valid / kedaluwarsa / tidak ditemukan)
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [nama_pemasok, nomor_sertifikat]
              properties:
                nama_pemasok: { type: string }
                nomor_sertifikat: { type: string }
                berlaku_sampai: { type: string, format: date, nullable: true }
                media_id: { type: integer, nullable: true }
                untuk_bahan: { type: array, items: { type: string } }
      responses: { '201': { description: "{id, status}" } }
...